requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.5.0
opencv-python==4.8.0.74
streamlit==1.27.0
